RE-173 Pin kibana version
The apt artifacts contain much later versions of
kibana in order to prepare for the implementation
of a much later release.

This patch pins the version installed in order to
get the implementation back to a working state
until the patch merges which updates the roles to
work with the new version.